Uploaded Steve's voice
Code was ripped literally from the example provided by the PyAudio documentation. I'll be implementing an array and some rng possibly in this function but I'll most likely implement that in a separate function. Steve has a pool of responses to "pick" from but for now I'll just make this function play the audio I think...